requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
# These ones added manually:
fuzzywuzzy==0.18.0
rapidfuzz==3.5.2
levenshtein==0.23.0
python-levenshtein==0.23.0
python-dotenv==1.0.0
azure-identity==1.11.0
pre-commit
###################
azure-keyvault-secrets==4.6.0
# azure-cli==2.40.0
atomicwrites==1.4.0
    # via pytest
attrs==21.4.0
    # via pytest
black==22.3.0
    # via -r requirements.in
blis==0.4.1
    # via
    #   spacy
    #   thinc
catalogue==1.0.0
    # via spacy
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.0.4
    # via
    #   black
    #   flask
    #   nltk
colorama==0.4.4
    # via
    #   click
    #   pytest
cymem==2.0.6
    # via
    #   preshed
    #   spacy
    #   thinc
emoji==1.7.0
    # via -r requirements.in
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz
    # via -r requirements.in
flake8==4.0.1
    # via -r requirements.in
Flask==2.1.1
    # via -r requirements.in
idna==3.3
    # via requests
importlib-metadata==4.11.3
    # via flask
iniconfig==1.1.1
    # via pytest
itsdangerous==2.1.2
    # via
    #   -r requirements.in
    #   flask
jinja2==3.1.1
    # via flask
joblib==1.1.0
    # via
    #   -r requirements.in
    #   nltk
    #   scikit-learn
json2html==1.3.0
    # via -r requirements.in
langdetect==1.0.7
    # via spacy-langdetect
markupsafe==2.1.1
    # via jinja2
mccabe==0.6.1
    # via flake8
murmurhash==1.0.6
    # via
    #   preshed
    #   spacy
    #   thinc
mypy-extensions==0.4.3
    # via black
msal==1.18.0b1 # Changed
msal-extensions~=1.0.0 #Changed
nltk==3.7
    # via textblob
numpy==1.22.3
    # via
    #   -r requirements.in
    #   blis
    #   pandas
    #   scikit-learn
    #   scipy
    #   spacy
    #   thinc
    #   xgboost
packaging==21.3
    # via pytest
pandas==1.4.2
    # via -r requirements.in
pathspec==0.9.0
    # via black
plac==1.1.3
    # via
    #   spacy
    #   thinc
platformdirs==2.5.1
    # via black
pluggy==1.0.0
    # via pytest
preshed==3.0.6
    # via
    #   spacy
    #   thinc
py==1.11.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pyopenssl<22.0.0
    # i did this
pyparsing==3.0.7
    # via packaging
pytest==7.1.1
    # via
    #   -r requirements.in
    #   spacy-langdetect
python-dateutil==2.8.2
    # via pandas
pytz==2022.1
    # via pandas
regex==2022.3.15
    # via
    #   -r requirements.in
    #   nltk
requests==2.27.1
    # via spacy
scikit-learn==1.0.2
    # via -r requirements.in
scipy==1.8.0
    # via
    #   scikit-learn
    #   xgboost
six==1.16.0
    # via
    #   langdetect
    #   python-dateutil
spacy==2.2.3
    # via
    #   -r requirements.in
    #   en-core-web-sm
spacy-langdetect==0.1.2
    # via -r requirements.in
srsly==1.0.5
    # via
    #   spacy
    #   thinc
textblob==0.17.1
    # via -r requirements.in
thinc==7.3.1
    # via spacy
threadpoolctl==3.1.0
    # via scikit-learn
tomli==2.0.1
    # via
    #   black
    #   pytest
tqdm==4.64.0
    # via
    #   nltk
    #   thinc
typing-extensions==4.1.1
    # via black
urllib3==1.26.7
    # via requests
wasabi==0.9.1
    # via
    #   spacy
    #   thinc
werkzeug==2.1.1
    # via flask
xgboost==1.5.2
    # via -r requirements.in
zipp==3.8.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
